Refactor MCP generate rules from base tool
