+++ crates/rulemorph_mcp/src/tools/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod analyze_input;
+pub(crate) mod generate_rules_from_base;
 pub(crate) mod list_ops;
